Connector authentication failure should not trigger login
Change the behaviour of the auth action to treat an authentication failure to an external service via a connector in application code as a server error and not an auth failure requiring the user to login again.

This situation is unlikely to happen unless the services are mis-configured or there are some infrastructure issues, since if the user is authenticated and authorized to the frontend app, they should also have access to any downstream APIs. However, this behaviour is correct should this occur, as forcing the user to login again will likely send them round in circles.