VPD-1035 prevent prod deployment warnings for the exit service
Warnings are generated about the exit survey url containing localhost before deployment to production. This is because the application.conf data is not overwritten by default. Change to use a url configuration item rather than a Service so that in app-config-base it can be configured as empty. This is the recommended configuration and will work in all environments downstream of local.